news section design changed
Design of news section has been changed.
Picture "about-me" has been added on a main page. This picture position needs to be corrected (flows little bit on right now).
Under this picture should be added the link "read about me" or something like that, that will lead us to the subpage about-me.

In the news section:
play button should play a sample of meditation OR lead to the subpage where it's possible to play a sample.

Under the part with the names of instruments shoud be a link "read-more" or something, that will lead to the subpage with a deeper instruments description.

Then think what more we can put on a main-page.. Maybe a video from the concert? <- this is to think.

If this is done, continue to fill and position the subpage about-me.